Introduce TileEntity GC (GarbageCollection) Flag
TL; DR Lets server know if tile entity in an unloaded chunk is being processed.

Allowing a tile entity to be processed in an unloaded chunk introduces a whole host of problems. When a chunk's onChunkUnload() is called, it adds its tile entities to a list of tile entities to be unloaded by the World class. However, when a tile entity's onUnload() is called, it can try to access tile entities that have already been unloaded, making the server try to reload them when they shouldn't be.

When I get a chance I will actually add usage for this in the Chunk and World classes.

The implementation of this flag has been tested on a server of over 30 players for nearly a week now with no problems whatsoever. It has actually reduced lag significantly.