Update libraries to latest versions
- A few weren't updated due to maven issues, but that'll be resolved in
the future.

I've barely tested this so I'm not 100% sure if it'll cause issues or
not in the future. All I did was just start up the server, play around
with some TNT (didn't install any mods) and ran the TPS command a bunch.
I did notice that the world ms went down by 12% though.

Known Issues:
Strange ^M in the console in front of every line.
Nothing Else.